+++ model_training/__init__.py
@@ -1,0 +1,2 @@
+from .get_prediction import Get_prediction
+from .model_training import BERT_Pretrained_Model, tokenizer
