--- src/backend/main.py
@@ -1,37 +1,0 @@
-import uuid
-
-import cv2
-import uvicorn
-from fastapi import File
-from fastapi import FastAPI
-from fastapi import UploadFile
-import numpy as np
-from PIL import Image
-
-import config
-import inference
-
-"""
-This is our server. FastAPI creates two endpoints, one dummy ("/") and one for serving our prediction ("/{style}"). The serving endpoint takes in a name as a URL parameter. We're using nine different trained models to perform style transfer, so the path parameter will tell us which model to choose. The image is accepted as a file over a POST request and sent to the inference function. Once the inference is complete, the file is stored on the local filesystem and the path is sent as a response.
-"""
-
-app = FastAPI()
-
-
-@app.get("/")
-def read_root():
-    return {"message": "Welcome from the API"}
-
-
-@app.post("/{style}")
-def get_image(style: str, file: UploadFile = File(...)):
-    image = np.array(Image.open(file.file))
-    model = config.STYLES[style]
-    output, resized = inference.inference(model, image)
-    name = f"/storage/{str(uuid.uuid4())}.jpg"
-    cv2.imwrite(name, output)
-    return {"name": name}
-
-
-if __name__ == "__main__":
-    uvicorn.run("main:app", host="0.0.0.0", port=8080)
