--- src/scripts/test.py
@@ -1,39 +1,0 @@
-from sklearn.metrics import confusion_matrix
-import matplotlib.pyplot as plt
-import numpy as np
-
-# Generate some sample data
-y_true = np.array([0, 1, 2, 0, 1, 2, 0, 1, 2])
-y_pred = np.array([0, 1, 1, 0, 2, 1, 0, 1, 2])
-
-# Compute confusion matrix
-cm = confusion_matrix(y_true, y_pred)
-
-# Define class labels
-class_names = ['Class 0', 'Class 1', 'Class 2']
-
-# Plot confusion matrix
-fig, ax = plt.subplots()
-im = ax.imshow(cm, interpolation='nearest', cmap=plt.cm.Blues)
-ax.figure.colorbar(im, ax=ax)
-ax.set(xticks=np.arange(cm.shape[1]),
-       yticks=np.arange(cm.shape[0]),
-       xticklabels=class_names, yticklabels=class_names,
-       title='Confusion matrix',
-       ylabel='True label',
-       xlabel='Predicted label')
-
-# Rotate the tick labels and set their alignment
-plt.setp(ax.get_xticklabels(), rotation=45, ha="right",
-         rotation_mode="anchor")
-
-# Loop over data dimensions and create text annotations
-thresh = cm.max() / 2.
-for i in range(cm.shape[0]):
-    for j in range(cm.shape[1]):
-        ax.text(j, i, format(cm[i, j], 'd'),
-                ha="center", va="center",
-                color="white" if cm[i, j] > thresh else "black")
-
-fig.tight_layout()
-plt.show()
