+++ visualization/backend/requirements.txt
@@ -1,0 +1,9 @@
+uvicorn == 0.20.0
+fastapi == 0.92.0
+torch == 1.13.1
+transformers == 4.25.1
+tqdm == 4.64.1
+pandas
+numpy
+scikit-learn
+python-multipart
