+++ visualization/frontend/pages/3_About us.py
@@ -1,0 +1,19 @@
+import streamlit as st
+
+st.write("Professor Loïc has assigned a school project that involves using Deep Learning to undertake, \
+    one of \
+    we choose the task 'sentiment analysis of movie reviews'. The project will require students to \
+        gather a dataset of movie reviews and their associated sentiments (positive, negative, or \
+        neutral) and then preprocess the data to feed it into the Bert model. The Bert model \
+        will be fine-tuned on the dataset and evaluated using various metrics such as \
+        accuracy, precision, recall, and F1-score. Finally, the model's performance \
+        will be visualized through confusion matrices and ROC curves. The project \
+        aims to give students an understanding of how Deep Learning models can \
+        be used for Natural Language Processing tasks such as sentiment analysis. \
+        By the end of the project, students will have developed skills in data preprocessing, \
+        model fine-tuning, and performance evaluation using the Bert model.")
+
+st.write("l'objectif")
+
+st.write("Les données utilisées (origine, format, statut juridique) et les traitements opérés sur celles-ci")
+
