+++ visualization/frontend/requirements.txt
@@ -1,0 +1,7 @@
+tqdm == 4.64.1
+pandas
+numpy
+sklearn
+streamlit == 1.12.0
+protobuf==3.20.1
+plotly
